+++ src/Icarus.Service/Interfaces/Categories/ICategoryService.cs
@@ -5,9 +5,9 @@ namespace Icarus.Service.Interfaces.Categories;
 
 public interface ICategoryService
 {
-    Task<bool> RemoveAsync(short id);
-    Task<CategoryForResultDto> RetrieveByIdAsync(short id);
+    Task<bool> RemoveAsync(long id);
+    Task<CategoryForResultDto> RetrieveByIdAsync(long id);
     Task<CategoryForResultDto> CreateAsync(CategoryForCreationDto dto);
-    Task<CategoryForResultDto> ModifyAsync(short id, CategoryForUpdateDto dto);
+    Task<CategoryForResultDto> ModifyAsync(long id, CategoryForUpdateDto dto);
     Task<IEnumerable<CategoryForResultDto>> RetrieveAllAsync(PaginationParams @params);
 }
